Add a package object, move some typed stuff to dir
This should be a no-op functionally (complain if you see anything
other than moving code or changing imports)
